# Conflicts:
#	src/tui/deep_research/engineered_loop_direct_review_tests.rs
#	src/tui/deep_research/engineered_loop_integration_tests.rs
#	src/tui/deep_research/engineered_loop_search_fallback_tests.rs
#	src/tui/deep_research/engineered_loop_tests.rs
#	src/tui/deep_research/prompts.rs
#	src/tui/deep_research/workflow/collection.js
#	src/tui/deep_research/workflow/direct_collection.js
#	src/tui/deep_research/workflow/loop.js
#	src/tui/deep_research/workflow/policy.js
#	src/tui/deep_research/workflow/runtime.js

+++ src/tui/panels/system/terminal.rs
@@ -106,9 +106,7 @@ mod tests {
         assert!(plain.contains("OSC 8 links"), "{plain}");
         assert!(plain.contains("OSC 52 copy"), "{plain}");
         assert!(
-            rendered
-                .lines()
-                .all(|line| visible_len(line) <= width as usize),
+            rendered.lines().all(|line| visible_len(line) <= width),
             "{rendered:?}"
         );
     }
